I found that my monitors were going red when the value was below the data-green threshold.
For example val = 100, data-green=500 and data-yellow=1000. This resulted in a red background and not a green one as you would expect.

The issue is that the values for data-green and data-yellow that are returned by

```javascript

document.getElementById(this.id).getAttribute('data-green')

```

are String and not numbers. When compared then "500" is in fact greater than "1000". I also found that "100" is less than "1000" because the string is shorter, not because the number the is less.

Parsing the numbers to ints fixed the issue for me.